# public/swagger.yaml
basePath: /api/v1
definitions:
  city.CitiesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/city.City'
        type: array
      meta:
        $ref: '#/definitions/models.PaginationMeta'
    type: object
  city.City:
    properties:
      country_code:
        type: string
      id:
        type: string
      name:
        type: string
      population:
        example: 3000000
        type: integer
      region_code:
        type: string
    type: object
  models.PaginationMeta:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
host: localhost:3020
info:
  contact:
    email: gotemplate-app@my-app.com
    name: gotemplate-app Support
    url: http://www.swagger.io/support
  description: A well-rounded template with everything you need to get started in
    golang, htmx and templ
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: gotemplate-app
  version: "1.0"
paths:
  /countries/{country_code}/cities:
    get:
      consumes:
      - application/json
      description: Get a list of cities with optional filtering including radius search
      parameters:
      - description: 'Country code - Example: AD, CM, ES, GA, US, CN, RU, FR, DE,
          etc...'
        in: path
        name: country_code
        required: true
        type: string
      - description: City name (fuzzy search)
        in: query
        name: name
        type: string
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/city.CitiesResponse'
      summary: Get cities
  /countries/{country_code}/cities/{id}:
    get:
      consumes:
      - application/json
      description: Get a city by its ID
      parameters:
      - description: City ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Country code - Example: AD, CM, ES, GA, US, CN, RU, FR, DE,
          etc...'
        in: path
        name: country_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/city.City'
      summary: Get a city by ID
swagger: "2.0"